refactor : 일부 출력 메세지 리팩토링

+++ src/main/java/lottogame/domain/Rank.java
@@ -14,7 +14,7 @@ public enum Rank {
     private static final int MAX_MATCH_NUMBER = 6;
 
     private int numberOfMatch;
-    private int prize;
+    private long prize;
 
     Rank(int numberOfMatch, int prize) {
         this.numberOfMatch = numberOfMatch;
